fix: avoid 307 redirect on POST session and set SameSite None in production
