--- .genkit/traces_idx/genkit.metadata
@@ -1,1 +1,0 @@
-{"version":"1.21.0"}
