+++ backend-hormonia/app/services/flow/errors/__init__.py
@@ -1,0 +1,14 @@
+"""Error handling utilities for Flow Services (QW-021)."""
+
+from .handler import FlowErrorHandler
+from .recovery import FlowRecoveryStrategy, RecoveryContext
+from .retry import RetryPolicy
+from .circuit_breaker import CircuitBreaker
+
+__all__ = [
+    "FlowErrorHandler",
+    "FlowRecoveryStrategy",
+    "RecoveryContext",
+    "RetryPolicy",
+    "CircuitBreaker",
+]
