+++ backend-hormonia/app/services/flow/execution/__init__.py
@@ -1,0 +1,1 @@
+"""Execution layer primitives for Flow Services (QW-021)."""
