+++ backend-hormonia/app/services/flow/monitoring/__init__.py
@@ -1,0 +1,1 @@
+"""Monitoring facade over flow analytics components."""
