+++ backend-hormonia/app/services/flow/monitoring/analytics.py
@@ -1,0 +1,6 @@
+"""Monitoring facade over the analytics module."""
+
+from ..analytics.analytics import FlowAnalytics
+from ..analytics import get_flow_analytics
+
+__all__ = ["FlowAnalytics", "get_flow_analytics"]
