+++ backend-hormonia/app/services/flow/monitoring/metrics.py
@@ -1,0 +1,5 @@
+"""Re-export metrics collector for clarity."""
+
+from ..analytics.metrics_collector import FlowMetricsCollector
+
+__all__ = ["FlowMetricsCollector"]
