+++ backend-hormonia/app/services/flow/validation/__init__.py
@@ -1,0 +1,7 @@
+"""Validation layer for Flow Services (QW-021)."""
+
+from .validator import FlowValidator
+from .integrity import FlowIntegrityChecker, IntegrityResult
+from .rules import ValidationRule
+
+__all__ = ["FlowValidator", "FlowIntegrityChecker", "IntegrityResult", "ValidationRule"]
